# BicycleRent.WebApplication/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Bicycle Rent API",
    "version": "v1"
  },
  "paths": {
...
    "/api/BicycleTypeDb": {
      "get": {
        "tags": [
          "BicycleTypeDb"
        ],
        "summary": "Получить все типы велосипедов",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BicycleTypeDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BicycleTypeDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BicycleTypeDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BicycleTypeDb"
        ],
        "summary": "Добавить тип велосипеда",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BicycleTypeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BicycleTypeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BicycleTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BicycleTypeDb/{id}": {
      "get": {
        "tags": [
          "BicycleTypeDb"
        ],
        "summary": "Получить тип велосипеда по id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BicycleTypeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BicycleTypeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BicycleTypeDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "BicycleTypeDb"
        ],
        "summary": "Обновить тип велосипеда по id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BicycleTypeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BicycleTypeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BicycleTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "BicycleTypeDb"
        ],
        "summary": "Удалить тип велосипеда по id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/RentDb": {
      "get": {
        "tags": [
          "RentDb"
        ],
        "summary": "Получить все арендны велосипедов",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RentDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RentDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RentDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RentDb"
        ],
        "summary": "Добавить аренду велосипеда",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RentDb/{id}": {
      "get": {
        "tags": [
          "RentDb"
        ],
        "summary": "Получить аренду велосипеда по id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RentDb"
        ],
        "summary": "Обновить аренду велосипеда по id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "RentDb"
        ],
        "summary": "Удалить аренду велосипеда по id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BicycleTypeDto": {
        "type": "object",
        "properties": {
          "typeId": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "RentDto": {
        "type": "object",
        "properties": {
          "rentId": {
            "type": "integer",
            "format": "int32"
          },
          "renterId": {
            "type": "integer",
            "format": "int32"
          },
          "bicycleId": {
            "type": "integer",
            "format": "int32"
          },
          "timeStart": {
            "type": "string",
            "format": "date-time"
          },
          "timeEnd": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}